% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMFOLL-package.R
\docType{data}
\name{atuaria}
\alias{atuaria}
\title{Atuaria data}
\format{
A data frame with 280 rows
}
\usage{
data(atuaria)
}
\description{
Time data, the data set refers to the time of retired women with temporary disabilities incorporated into the Mexican public insurance system until they died in 2004, corresponding to 280 lives.
}
\details{
The variable is as follow:

\itemize{
\item x Variable of time
}
}
\keyword{datasets}
