% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaburrxii.R
\name{BetaBurrXII}
\alias{BetaBurrXII}
\alias{dbbxii}
\alias{pbbxii}
\alias{qbbxii}
\alias{rbbxii}
\title{The Beta Burr XII Distribution.}
\usage{
dbbxii(param, x, log = FALSE)

pbbxii(param, q, lower.tail = TRUE, log = FALSE)

qbbxii(param, p, lower.tail = TRUE, log = FALSE)

rbbxii(param, n)
}
\arguments{
\item{param}{Vector with five parameters, first scale parameter, the other ones are shape parameters.}

\item{x, q}{Vector of quantiles.}

\item{log}{Logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}
}
\value{
dbbxii gives the density, pbbxii gives the distribution function, qbbxii quantile function, rbbxii random generation function.
}
\description{
Provides density, distribution function, quantile function and random generation for the Beta Burr XII distribution.
}
\examples{
pbbxii(c(1,1,1,1,1),5)

dbbxii(c(5,5,5,5,5),c(2.1,3,2.5))
}
