% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burrxii.R
\name{BXII}
\alias{BXII}
\alias{dbxii}
\alias{pbxii}
\alias{qbxii}
\alias{rbxii}
\title{The Burr XII Distribution.}
\usage{
dbxii(param, x, log = FALSE)

pbxii(param, q, lower.tail = TRUE, log = FALSE)

qbxii(param, p, lower.tail = TRUE, log = FALSE)

rbxii(param, n)
}
\arguments{
\item{param}{Vector with three parameters, first the scale parameter and then the two shape parameters.}

\item{x, q}{Vector of quantiles.}

\item{log}{Logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}
}
\value{
dbxii gives the density, pbxii gives the distribution function, qbxii quantile function, rbxii random generation function.
}
\description{
Provides density, distribution function, quantile function and random generation for the Burr XII distribution.
}
\examples{
pbxii(c(4,2.1,3),5)

dbxii(c(2,3,3),c(2.1,1.2,0.6))

x <- rbxii(c(1,1,1),1e+3)
plot(ecdf(x), lwd = 3)
curve(pbxii(c(1,1,1),x), add = TRUE, col = 2, from = 0, lwd = 2)
legend('right', legend = c('Random', 'Theoretical'), lty = 1, col = 1:2, bty = 'n', lwd = 2)

}
