% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMFOLL-package.R
\docType{data}
\name{concrete}
\alias{concrete}
\title{Concrete data}
\format{
A data frame with 14 rows
}
\usage{
data(concrete)
}
\description{
Time data
}
\details{
The variable is as follow:

\itemize{
\item x Variable of time
}
}
\keyword{datasets}
