% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descritiva.R
\name{descritiva}
\alias{descritiva}
\title{descritiva - Calculation of some descriptive statistics.}
\usage{
descritiva(
  x,
  colb = "blue",
  colh = "green",
  xlab = "values of the dataset",
  breaks = 40,
  main = "",
  cens = FALSE,
  b = NULL,
  ...
)
}
\arguments{
\item{x, cens}{A numeric vector.}

\item{colb}{Color box-plot.}

\item{colh}{Color histogram.}

\item{xlab}{Axis label.}

\item{breaks}{Values of the breaks.}
}
\value{
Some descriptive statistics of a numeric vector
}
\description{
The function descritiva shows some descriptive statistics of a numeric vector
}
\examples{

data(atuaria)
descritiva(atuaria$x)

}
\seealso{
\url{http://www.r-project.org}
}
