% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechet.R
\name{Frechet}
\alias{Frechet}
\alias{dfr}
\alias{pfr}
\alias{qfr}
\alias{rfr}
\title{The Frechet Distribution.}
\usage{
dfr(param, x, log = FALSE)

pfr(param, q, lower.tail = TRUE, log = FALSE)

qfr(param, p, lower.tail = TRUE, log = FALSE)

rfr(param, n)
}
\arguments{
\item{param}{Vector with shape parameter first and scale parameter second.}

\item{x, q}{Vector of quantiles.}

\item{log}{Logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}
}
\value{
dfr gives the density, pfr gives the distribution function, qfr quantile function, rfr random generation function.
}
\description{
Provides density, distribution function, quantile function and random generation for the Frechet distribution.
}
\examples{
pfr(c(4,2.1,3),5)

dfr(c(2,3,3),c(2.1,1.2,0.6))
}
