% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kumaraswamyburrxii.R
\name{KwBXII}
\alias{KwBXII}
\alias{dkwbxii}
\alias{pkwbxii}
\alias{qkwbxii}
\alias{rkwbxii}
\title{The Kumaraswamy Burr XII Distribution.}
\usage{
dkwbxii(param, x, log = FALSE)

pkwbxii(param, q, lower.tail = TRUE, log = FALSE)

qkwbxii(param, p, lower.tail = TRUE, log = FALSE)

rkwbxii(param, n)
}
\arguments{
\item{param}{Vector with five parameters, first scale parameter, the other ones are shape parameters.}

\item{x}{Vector of quantiles.}
}
\value{
dkwbxii gives the density, pkwbxii gives the distribution function, qkwbxii quantile function, rkwbxii random generation function.
}
\description{
Provides density, distribution function, quantile function and random generation for the Kumaraswamy Burr XII distribution.
}
\examples{
pkwbxii(c(1,1,1,1,1),5)

dkwbxii(c(2,2,2,2,2),c(2.1,1.2,0.6))
}
