% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglogistica.R
\name{LogLogistic}
\alias{LogLogistic}
\alias{dll}
\alias{pll}
\alias{qll}
\alias{rll}
\title{The Log-Logistic Distribution.}
\usage{
dll(param, x, log = FALSE)

pll(param, q, lower.tail = TRUE, log = FALSE)

qll(param, p, lower.tail = TRUE, log = FALSE)

rll(param, n)
}
\arguments{
\item{param}{Vector with scale parameter first and shape parameter second.}

\item{x, q}{Vector of quantiles.}

\item{log}{Logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}
}
\value{
dll gives the density, pll gives the distribution function, qll quantile function, rll random generation function.
}
\description{
Provides density, distribution function, quantile function and random generation for the Log-Logistic distribution.
}
\examples{
pll(c(4,2.1),5)

dll(c(2,3),c(2.1,1.2,0.6))
}
