% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMFOLL-package.R
\docType{data}
\name{melanoma}
\alias{melanoma}
\title{Melanoma data}
\format{
A data frame with 417 rows
}
\usage{
data(melanoma)
}
\description{
Time data, the data are on cutaneous melanoma (a type of malignant cancer) that assesses the performance of a post-operative treatment with a high dose of the drug interferon alfa-2b to prevent a recurrence. Patients were included in the study from 1991 to 1995 and follow-up was carried out through 1998. Survival time X is referred to as the time to patient death.
}
\details{
The variable is as follow:

\itemize{
\item x Variable of time
\item cens Censure
\item x0 Covariate
\item x1 Covariate
\item x2 Covariate
\item x3 Covariate
\item x4 Covariate
\item x5 Covariate
\item x6 Covariate
\item obs Covariate
}
}
\keyword{datasets}
