% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollbxii.R
\name{OLLBXII}
\alias{OLLBXII}
\alias{dollbxii}
\alias{pollbxii}
\alias{qollbxii}
\alias{rollbxii}
\title{The Odd Log-Logistic Burr XII Distribution.}
\usage{
dollbxii(param, x, log = FALSE)

pollbxii(param, q, lower.tail = TRUE, log = FALSE)

qollbxii(param, p, lower.tail = TRUE, log = FALSE)

rollbxii(param, n)
}
\arguments{
\item{param}{Vector with four parameters, first the scale parameter, then the two shape parameters and the last one the odd parameter.}

\item{x}{Vector of quantiles.}
}
\value{
dollbxii gives the density, pollbxii gives the distribution function, qollbxii quantile function, rollbxii random generation function.
}
\description{
Provides density, distribution function, quantile function and random generation for the Odd Log-Logistic Burr XII distribution.
}
\examples{
pollbxii(c(4,2.1,3,4),c(4,5))

dollbxii(c(2,3,3,3),c(2.1,1.2,0.6))
}
