% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollexponencial.R
\name{OLLexpo}
\alias{OLLexpo}
\alias{dollexpo}
\alias{pollexpo}
\alias{qollexpo}
\alias{rollexpo}
\title{The Odd Log-Logistic exponential Distribution.}
\usage{
dollexpo(param, x, log = FALSE)

pollexpo(param, q, lower.tail = TRUE, log = FALSE)

qollexpo(param, p, lower.tail = TRUE, log = FALSE)

rollexpo(param, n)
}
\arguments{
\item{param}{Vector with two parameters, first scale parameter and second the odd parameter.}

\item{x, q}{Vector of quantiles.}

\item{log}{Logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}
}
\value{
dollexpo gives the density, pollexpo gives the distribution function, qollexpo quantile function, rollexpo random generation function.
}
\description{
Provides density, distribution function, quantile function and random generation for the Odd Log-Logistic exponential distribution.
}
\examples{
pollexpo(c(1,2),c(4,0.6))

dollexpo(c(1,2),c(4,0.6))
}
