% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollfit.R
\name{OllFit}
\alias{OllFit}
\alias{oll.fit}
\title{Fit the data with Odd Log-Logistic distributions}
\usage{
oll.fit(
  dist,
  starts,
  data,
  method = "B",
  domain = c(0, Inf),
  mle = NULL,
  cens = NULL
)
}
\arguments{
\item{dist}{The name of the studied distributions may be 'fr', 'bxii', 'wei', 'll', 'expo', 'norm', 'gamma', 'ln', 'hn', 'ollfr', 'ollbxii', 'ollwei', 'ollll', 'ollexpo', 'ollnorm', 'ollgamma', 'ollln', 'ollhn', 'bbxii' or 'kwbxii'.}

\item{starts}{A vector with initial gess.}

\item{data, cens}{The data set.}

\item{method}{The method to be used, by default 'B'.}

\item{...}{}
}
\value{
Fit data of a distribution.
}
\description{
The function oll.fit shows some fitted results
}
\examples{
data(atuaria)
x=atuaria$x
oll.fit(dist = 'fr',starts = c(1,1),data = x)

data(melanoma)
cens=melanoma$cens
x=melanoma$x
oll.fit(dist = 'ollwei',starts = c(1,1,1),x, cens = cens)
}
\seealso{
\url{http://www.r-project.org}
}
