% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollfrechet.R
\name{OLLFrechet}
\alias{OLLFrechet}
\alias{dollfr}
\alias{pollfr}
\alias{qollfr}
\alias{rollfr}
\title{The Odd Log-Logistic Frechet Distribution.}
\usage{
dollfr(param, x, log = FALSE)

pollfr(param, q, lower.tail = TRUE, log = FALSE)

qollfr(param, p, lower.tail = TRUE, log = FALSE)

rollfr(param, n)
}
\arguments{
\item{param}{Vector with three parameters, first shape parameter, second scale parameter and third odd parameter.}

\item{x, q}{Vector of quantiles.}

\item{log}{Logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}
}
\value{
dollfr gives the density, pollfr gives the distribution function, qollfr quantile function, rollfr random generation function.
}
\description{
Provides density, distribution function, quantile function and random generation for the Odd log-logistic Frechet distribution.
}
\examples{
pollfr(c(2,2,3),c(5.1,5.6))

dollfr(c(1,2,3),c(4,0.6))

x <- rollfr(c(.91,6,2),1e+4)
plot(ecdf(x), lwd = 3)
curve(pollfr(c(.91,6,2),x), add = TRUE, col = 2, from = 0, lwd = 2)
legend('right', legend = c('Random', 'Theoretical'), lty = 1, col = 1:2, bty = 'n', lwd = 2)
}
