% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollgama.R
\name{OLLgamma}
\alias{OLLgamma}
\alias{dollgamma}
\alias{pollgamma}
\alias{qollgamma}
\alias{rollgamma}
\title{The Odd Log-Logistic gamma Distribution.}
\usage{
dollgamma(param, x, log = FALSE)

pollgamma(param, q, lower.tail = TRUE, log = FALSE)

qollgamma(param, p, lower.tail = TRUE, log = FALSE)

rollgamma(param, n)
}
\arguments{
\item{param}{Vector with three parameters, first shape parameter, second scale parameter and third odd parameter.}

\item{x, q}{Vector of quantiles.}

\item{log}{Logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}
}
\value{
dollgamma gives the density, pollgamma gives the distribution function, qollgamma quantile function, rollgamma random generation function.
}
\description{
Provides density, distribution function, quantile function and random generation for the Odd Log-Logistic gamma distribution.
}
\examples{
pollgamma(c(1,2,3),c(4,0.6))

dollgamma(c(1,2,3),c(4,0.6))
}
