% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollhalfnormal.R
\name{OLLHN}
\alias{OLLHN}
\alias{dollhn}
\alias{pollhn}
\alias{qollhn}
\alias{rollhn}
\title{The Odd Log-Logistic Half-Normal Distribution.}
\usage{
dollhn(param, x, log = FALSE)

pollhn(param, q, lower.tail = TRUE, log = FALSE)

qollhn(param, p, lower.tail = TRUE, log = FALSE)

rollhn(param, n)
}
\arguments{
\item{param}{Vector with three parameters, first shape parameter, second scale parameter and third odd parameter.}

\item{x, q}{Vector of quantiles.}

\item{log}{Logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}
}
\value{
dollhn gives the density, pollhn gives the distribution function, qollhn quantile function, rollhn random generation function.
}
\description{
Provides density, distribution function, quantile function and random generation for the Odd Log-Logistic Half-Normal distribution.
}
\examples{
pollhn(c(1,2,3),c(4,0.6))

dollhn(c(1,2,3),c(4,0.6))
}
