% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollloglogistica.R
\name{OLLLL}
\alias{OLLLL}
\alias{dollll}
\alias{pollll}
\alias{qollll}
\alias{rollll}
\title{The Odd Log-Logistic Log-Logistic Distribution.}
\usage{
dollll(param, x, log = FALSE)

pollll(param, q, lower.tail = TRUE, log = FALSE)

qollll(param, p, lower.tail = TRUE, log = FALSE)

rollll(param, n)
}
\arguments{
\item{param}{Vector with three parameters, first scale parameter, second shape parameter and third odd parameter.}

\item{x, q}{Vector of quantiles.}

\item{log}{Logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}
}
\value{
dollll gives the density, pollll gives the distribution function, qollll quantile function, rollll random generation function.
}
\description{
Provides density, distribution function, quantile function and random generation for the Odd Log-Logistic Log-Logistic distribution.
}
\examples{
pollll(c(1,2,3),c(4,0.6))

dollll(c(1,2,3),c(4,0.6))
}
