% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olllognormal.R
\name{OLLLN}
\alias{OLLLN}
\alias{dollln}
\alias{pollln}
\alias{qollln}
\alias{rollln}
\title{The Odd Log-Logistic Log-Normal Distribution.}
\usage{
dollln(param, x, log = FALSE)

pollln(param, q, lower.tail = TRUE, log = FALSE)

qollln(param, p, lower.tail = TRUE, log = FALSE)

rollln(param, n)
}
\arguments{
\item{param}{Vector with three parameters, first scale parameter, second shape parameter and third odd parameter.}

\item{x, q}{Vector of quantiles.}

\item{log}{Logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}
}
\value{
dollln gives the density, pollln gives the distribution function, qollln quantile function, rollln random generation function.
}
\description{
Provides density, distribution function, quantile function and random generation for the Odd Log-Logistic Log-Normal distribution.
}
\examples{
pollln(c(1,2,3),c(4,0.6))

dollln(c(1,2,3),c(4,0.6))
}
