% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollnormal.R
\name{OLLnorm}
\alias{OLLnorm}
\alias{dollnorm}
\alias{pollnorm}
\alias{qollnorm}
\alias{rollnorm}
\title{The Odd Log-Logistic normal Distribution.}
\usage{
dollnorm(param, x, log = FALSE)

pollnorm(param, q, lower.tail = TRUE, log = FALSE)

qollnorm(param, p, lower.tail = TRUE, log = FALSE)

rollnorm(param, n)
}
\arguments{
\item{param}{Vector with three parameters, first scale parameter, second shape parameter and third odd parameter.}

\item{x, q}{Vector of quantiles.}

\item{log}{Logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}
}
\value{
dollnorm gives the density, pollnorm gives the distribution function, qollnorm quantile function, rollnorm random generation function.
}
\description{
Provides density, distribution function, quantile function and random generation for the Odd Log-Logistic normal distribution.
}
\examples{
pollnorm(c(1,2,3),c(4,0.6))

dollnorm(c(1,2,3),c(4,0.6))
}
