% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressao.R
\name{OllReg}
\alias{OllReg}
\alias{oll.reg}
\title{Regression to Odd Log-Logistic distributions}
\usage{
oll.reg(dist, starts, y, X, delta, domain = c(0, Inf), plot = T)
}
\arguments{
\item{dist}{The name of the studied distributions may be 'll', 'bxii', 'ollbxii' or 'ollghn'.}

\item{starts}{A vector with initial gess.}

\item{y, X, delta}{The data set.}

\item{...}{}
}
\value{
Fit data of a distribution.
}
\description{
The function oll.reg shows some regression results
}
\examples{
dados <- read.table("http://users.stat.ufl.edu/~winner/data/alc_elim.dat", sep = "",header = FALSE)
head(dados)
y <- (with(dados, V4))
delta <- rep(1, length(y))
x0 <- rep(1, length(y))
x1 <- (with(dados, V3))
x2 <- (with(dados, V2))
X = as.matrix(cbind(x0,x1,x2))

opt1 = oll.reg(dist = 'll',starts = c(1,1,1,1),y = y,X = X,delta = delta, plot = TRUE); opt1
opt2 = oll.reg(dist = 'bxii',starts = c(1,1,1,1,1),y = y,X = X,delta = delta, plot = TRUE); opt2
opt3 = oll.reg('ollbxii',c(opt2$par, .9),y,X,delta, plot = TRUE); opt3
}
\seealso{
\url{http://www.r-project.org}
}
