% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMFOLL-package.R
\docType{package}
\name{SMFOLL}
\alias{SMFOLL}
\title{SMFOLL: Solution to Modeling the Odd Log-Logistic Family.}
\description{
The SMFOLL package provides important
functions and datasets
for survival analysis.
Some distributions of the Odd Log-Logistic Family.
(mara.caroline.torres@uel.br)
}
\section{SMFOLL functions}{

\code{\link[SMFOLL]{BetaBurrXII}}

\code{\link[SMFOLL]{BXII}}

\code{\link[SMFOLL]{descritiva}}

\code{\link[SMFOLL]{Frechet}}

\code{\link[SMFOLL]{HalfNormal}}

\code{\link[SMFOLL]{KwBXII}}

\code{\link[SMFOLL]{LogLogistic}}

\code{\link[SMFOLL]{OLLBXII}}

\code{\link[SMFOLL]{OLLexpo}}

\code{\link[SMFOLL]{OllFit}}

\code{\link[SMFOLL]{OLLFrechet}}

\code{\link[SMFOLL]{OLLgamma}}

\code{\link[SMFOLL]{OLLHN}}

\code{\link[SMFOLL]{OLLLL}}

\code{\link[SMFOLL]{OLLLN}}

\code{\link[SMFOLL]{OLLnorm}}

\code{\link[SMFOLL]{OllReg}}

\code{\link[SMFOLL]{OLLWeibull}}

\code{\link[SMFOLL]{test}}

\code{\link[SMFOLL]{Weibull}}
}

