% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{Weibull}
\alias{Weibull}
\alias{dwei}
\alias{pwei}
\alias{qwei}
\alias{rwei}
\title{The Weibull Distribution.}
\usage{
dwei(param, x, log = FALSE)

pwei(param, q, lower.tail = TRUE, log = FALSE)

qwei(param, p, lower.tail = TRUE, log = FALSE)

rwei(param, n)
}
\arguments{
\item{param}{Vector with scale parameter first and shape parameter second.}

\item{x, q}{Vector of quantiles.}

\item{log}{Logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations.}
}
\value{
dwei gives the density, pwei gives the distribution function, qwei quantile function, rwei random generation function.
}
\description{
Provides density, distribution function, quantile function and random generation for the Weibull distribution.
}
\examples{

pwei(c(1,2),1)

data(atuaria)
pwei(c(0.4,0.2),atuaria)

dwei(c(0.1,0.99),0.6)

x <- rwei(c(1,1),1e+3)
plot(ecdf(x), lwd = 3)
curve(pweibull(x,1,1), add = TRUE, col = 2, from = 0, lwd = 2)
legend('right', legend = c('Random', 'Theoretical'), lty = 1, col = 1:2, bty = 'n', lwd = 2)
}
